Handle Django container stopping
Currently, ghostwriter-cli does not detect if if the Django container
throws an error when running commands like `install`, and waits forever
in such a case. This rewrites the waitForDjango function to check that
the container is still running, and returns an error if it stops without
listening for events, with a message guiding users to the Django
container logs.