Add pg-upgrade subcommand
PostgreSQL's data requires an upgrade step when changing major versions,
which requires both the old and new versions to be installed.

This automates upgrades the PostgreSQL data, by exporting the data using
the data's version (detected from the volume), clearing the volume, then
re-importing it with the version in the docker-compose file.